requirements.txt
# ─────────────────────────────────────────
#  Pro-Rooms – Python Dependencies
# ─────────────────────────────────────────

# Core Flask
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
MarkupSafe==3.0.2

# Database – PostgreSQL
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10        # PostgreSQL adapter (binary, no build required)

# Google OAuth / SSO
Authlib==1.3.2
requests==2.32.3

# Environment variables
python-dotenv==1.1.0